Add bounds_nav_l[on|at] to drop vars list
Some 201905 results files have extra NEMO variables called bounds_nav_lon
and bounds_nav_lat in addition to or in place of bounds_lon & bounds_lat.
If the *nav* pair are present in the dataset we are dumping to netCDF4,
the metadata cleanup fails. So, make sure they get dropped.