Add demo of kerchunk using json & parquet storage
Store per-netcdf kerchunk references as JSON so that they can be
reused because the effort to build them for a long time series
with multiple variable groups will be significant.
Store combined reference using Parquet because it requires
about 40% of the storage space of a combined JSON reference,
and it should have a smaller memory footprint.