Handle Pitt River & Fraser River proxy
Pitt River is truncated near the east domain boundary so that it is not
an open boundary.

Fraser River is also truncated near the east domain boundary, and
extended in the negative y direction to provide a channel to get
acceptable representation of tidal currents in the river.

The Fraser extension differs from previous bathymetries in that it
goes in the negative y direction instead of the positive. Also, in
previous bathymetries it actually exteded from the Pitt River, not the
Fraser.